+++ instrumentation/internal/config.go
@@ -24,18 +24,20 @@ import (
 	"go.opentelemetry.io/otel/propagation"
 	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 	"go.opentelemetry.io/otel/trace"
-
-	splunkotel "github.com/signalfx/splunk-otel-go"
 )
 
 // Config contains configuration options.
 type Config struct {
 	// instName is the name of the instrumentation this Config is used for.
 	instName string
+	// Version is the version of the instrumentation this Config is used for.
+	// It has to be set before user-provided options are applied.
+	Version string
 
-	Tracer           trace.Tracer
-	Meter            metric.Meter
-	Propagator       propagation.TextMapPropagator
+	Tracer     trace.Tracer
+	Meter      metric.Meter
+	Propagator propagation.TextMapPropagator
+
 	DefaultStartOpts []trace.SpanStartOption
 }
 
@@ -53,18 +55,10 @@ func NewConfig(instrumentationName string, options ...Option) *Config {
 	}
 
 	if c.Tracer == nil {
-		c.Tracer = otel.Tracer(
-			c.instName,
-			trace.WithInstrumentationVersion(splunkotel.Version()),
-			trace.WithSchemaURL(semconv.SchemaURL),
-		)
+		c.Tracer = c.tracer(otel.GetTracerProvider())
 	}
 	if c.Meter == nil {
-		c.Meter = global.Meter(
-			c.instName,
-			metric.WithInstrumentationVersion(splunkotel.Version()),
-			metric.WithSchemaURL(semconv.SchemaURL),
-		)
+		c.Meter = c.meter(global.MeterProvider())
 	}
 
 	if c.Propagator == nil {
@@ -97,11 +91,7 @@ func (c *Config) Copy() *Config {
 // from c is used.
 func (c *Config) ResolveTracer(ctx context.Context) trace.Tracer {
 	if span := trace.SpanFromContext(ctx); span.SpanContext().IsValid() {
-		return span.TracerProvider().Tracer(
-			c.instName,
-			trace.WithInstrumentationVersion(splunkotel.Version()),
-			trace.WithSchemaURL(semconv.SchemaURL),
-		)
+		return c.tracer(span.TracerProvider())
 	}
 	return c.Tracer
 }
@@ -142,4 +132,22 @@ func (c *Config) WithSpan(ctx context.Context, name string, f func(context.Conte
 	span.End()
 
 	return err
+}
+
+// tracer creates a tracer using the passed TracerProvider.
+func (c *Config) tracer(tp trace.TracerProvider) trace.Tracer {
+	opts := []trace.TracerOption{trace.WithSchemaURL(semconv.SchemaURL)}
+	if c.Version != "" {
+		opts = append(opts, trace.WithInstrumentationVersion(c.Version))
+	}
+	return tp.Tracer(c.instName, opts...)
+}
+
+// meter creates a meter using the passed MeterProvider.
+func (c *Config) meter(mp metric.MeterProvider) metric.Meter {
+	opts := []metric.MeterOption{metric.WithSchemaURL(semconv.SchemaURL)}
+	if c.Version != "" {
+		opts = append(opts, metric.WithInstrumentationVersion(c.Version))
+	}
+	return mp.Meter(c.instName, opts...)
 }