+++ instrumentation/internal/option.go
@@ -18,10 +18,7 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/propagation"
-	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 	"go.opentelemetry.io/otel/trace"
-
-	splunkotel "github.com/signalfx/splunk-otel-go"
 )
 
 // Option applies options to a configuration.
@@ -41,11 +38,7 @@ func (o OptionFunc) Apply(c *Config) {
 // a configuration.
 func WithTracerProvider(tp trace.TracerProvider) Option {
 	return OptionFunc(func(c *Config) {
-		c.Tracer = tp.Tracer(
-			c.instName,
-			trace.WithInstrumentationVersion(splunkotel.Version()),
-			trace.WithSchemaURL(semconv.SchemaURL),
-		)
+		c.Tracer = c.tracer(tp)
 	})
 }
 
@@ -53,11 +46,7 @@ func WithTracerProvider(tp trace.TracerProvider) Option {
 // a configuration.
 func WithMeterProvider(mp metric.MeterProvider) Option {
 	return OptionFunc(func(c *Config) {
-		c.Meter = mp.Meter(
-			c.instName,
-			metric.WithInstrumentationVersion(splunkotel.Version()),
-			metric.WithSchemaURL(semconv.SchemaURL),
-		)
+		c.Meter = c.meter(mp)
 	})
 }