+++ CHANGELOG.md
@@ -8,6 +8,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix `telemetry.sdk.version` resource attribute
+  to properly return the `github.com/signalfx/splunk-otel-go/distro` Go module version.
+  (#1990)
+- Fix instrumentation scope version value
+  to properly return the instrumentation library versions. (#1990)
+
 ## [1.5.0] - 2023-03-30
 
 ### Added